pyproject.toml
[tool.poetry]
name = "agendamento-medico"
version = "0.1.0"
description = ""
authors = ["Christian Lopes <christian.lima@ntendencia.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.35"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
gunicorn = "^23.0.0"
pydantic-settings = "^2.5.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
isort = "^5.13.2"
black = "^24.8.0"
flake8 = "^7.1.1"
mypy = "^1.11.2"
pre-commit = "^3.8.0"
pytest-random-order = "^1.1.1"
httpx = "^0.27.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
